refactor: fix package name
