+++ src/main/java/org/obd/metrics/codec/batch/encoder/AdjustableBatchSizeEncoder.java
@@ -14,7 +14,7 @@
  * express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.obd.metrics.codec.batch.enocder;
+package org.obd.metrics.codec.batch.encoder;
 
 import java.util.List;
